사용자가 전송한 평가 정보를 저장하는 기능 구현
장소에 대한 평가가 있는 평가정보에 만약 해당 장소에 대한 정보가 없으면
장소를 추가하고, 평가 데이터를 넣도록 구현

+++ src/main/java/kr/ac/hansung/maldives/web/model/Evaluation.java
@@ -23,15 +23,15 @@ public class Evaluation {
 				parameters=@Parameter(value="position", name="property"))
 	private Long positionIdx;
 	
-	private Double eva1;
+	private Float eva1;
 	
-	private Double eva2;
+	private Float eva2;
 	
-	private Double eva3;
+	private Float eva3;
 	
-	private Double eva4;
+	private Float eva4;
 	
-	private Double eva5;
+	private Float eva5;
 	
 	@OneToOne
 	@JoinColumn(name="position_idx")
